added surfaces and slightly improved the style of the code
