+++ README.md
@@ -1,7 +1,7 @@
 # raytracer
 A ray tracer written in pure C ++.
 
-The test.ppm file contains the drawn image.
+The /build/test.ppm file contains the drawn image.
 
 To build and run a project:
 mkdir -p build && cd build && cmake .. && make && ./raytracer
