--- image/IImage.cpp
@@ -1,1 +1,0 @@
-#include "IImage.hpp"
