--- raytracer/Camera.cpp
@@ -1,1 +1,0 @@
-#include "Camera.hpp"
